+++ src/fonduer/utils/logging/__init__.py
@@ -1,0 +1,3 @@
+from fonduer.utils.logging.tensorboard_writer import TensorBoardLogger
+
+__all__ = ["TensorBoardLogger"]