+++ tests/candidates/test_candidates.py
@@ -202,10 +202,7 @@ def test_cand_gen_cascading_delete(caplog):
         logger.info("Using two cores.")
         PARALLEL = 2  # Travis only gives 2 cores
 
-    def do_nothing_matcher(fig):
-        return True
-
-    max_docs = 10
+    max_docs = 1
     session = Meta.init("postgresql://localhost:5432/" + DB).Session()
 
     docs_path = "tests/data/html/"
@@ -219,7 +216,7 @@ def test_cand_gen_cascading_delete(caplog):
     )
     corpus_parser.apply(doc_preprocessor, parallelism=PARALLEL)
     assert session.query(Document).count() == max_docs
-    assert session.query(Sentence).count() == 5548
+    assert session.query(Sentence).count() == 799
     docs = session.query(Document).order_by(Document.name).all()
 
     # Mention Extraction
@@ -234,9 +231,9 @@ def test_cand_gen_cascading_delete(caplog):
     )
     mention_extractor.apply(docs, parallelism=PARALLEL)
 
-    assert session.query(Mention).count() == 370
-    assert session.query(Part).count() == 234
-    assert session.query(Temp).count() == 136
+    assert session.query(Mention).count() == 93
+    assert session.query(Part).count() == 70
+    assert session.query(Temp).count() == 23
     part = session.query(Part).order_by(Part.id).all()[0]
     temp = session.query(Temp).order_by(Temp.id).all()[0]
     logger.info(f"Part: {part.context}")
@@ -251,17 +248,17 @@ def test_cand_gen_cascading_delete(caplog):
 
     candidate_extractor.apply(docs, split=0, parallelism=PARALLEL)
 
-    assert session.query(PartTemp).count() == 3879
-    assert session.query(Candidate).count() == 3879
+    assert session.query(PartTemp).count() == 1432
+    assert session.query(Candidate).count() == 1432
     assert docs[0].name == "112823"
     assert len(docs[0].parts) == 70
-    assert len(docs[0].temps) == 24
+    assert len(docs[0].temps) == 23
 
     # Delete from parent class should cascade to child
     x = session.query(Candidate).first()
     session.query(Candidate).filter_by(id=x.id).delete(synchronize_session="fetch")
-    assert session.query(PartTemp).count() == 3878
-    assert session.query(Candidate).count() == 3878
+    assert session.query(Candidate).count() == 1431
+    assert session.query(PartTemp).count() == 1431
 
     # Clearing Mentions should also delete Candidates
     mention_extractor.clear()
@@ -286,7 +283,7 @@ def test_cand_gen(caplog):
     def do_nothing_matcher(fig):
         return True
 
-    max_docs = 10
+    max_docs = 1
     session = Meta.init("postgresql://localhost:5432/" + DB).Session()
 
     docs_path = "tests/data/html/"
@@ -300,7 +297,7 @@ def test_cand_gen(caplog):
     )
     corpus_parser.apply(doc_preprocessor, parallelism=PARALLEL)
     assert session.query(Document).count() == max_docs
-    assert session.query(Sentence).count() == 5548
+    assert session.query(Sentence).count() == 799
     docs = session.query(Document).order_by(Document.name).all()
 
     # Mention Extraction
@@ -339,10 +336,10 @@ def test_cand_gen(caplog):
     )
     mention_extractor.apply(docs, parallelism=PARALLEL)
 
-    assert session.query(Part).count() == 234
-    assert session.query(Volt).count() == 107
-    assert session.query(Temp).count() == 136
-    assert session.query(Fig).count() == 223
+    assert session.query(Part).count() == 70
+    assert session.query(Volt).count() == 33
+    assert session.query(Temp).count() == 23
+    assert session.query(Fig).count() == 31
     part = session.query(Part).order_by(Part.id).all()[0]
     volt = session.query(Volt).order_by(Volt.id).all()[0]
     temp = session.query(Temp).order_by(Temp.id).all()[0]
@@ -379,9 +376,9 @@ def test_cand_gen(caplog):
 
     candidate_extractor.apply(docs, split=0, parallelism=PARALLEL)
 
-    assert session.query(PartTemp).count() == 4141
-    assert session.query(PartVolt).count() == 3610
-    assert session.query(Candidate).count() == 7751
+    assert session.query(PartTemp).count() == 1610
+    assert session.query(PartVolt).count() == 2310
+    assert session.query(Candidate).count() == 3920
     candidate_extractor.clear_all(split=0)
     assert session.query(Candidate).count() == 0
     assert session.query(PartTemp).count() == 0
@@ -393,10 +390,9 @@ def test_cand_gen(caplog):
     )
 
     candidate_extractor.apply(docs, split=0, parallelism=PARALLEL)
-
-    assert session.query(PartTemp).count() == 3879
-    assert session.query(PartVolt).count() == 3610
-    assert session.query(Candidate).count() == 7489
+    assert session.query(PartTemp).count() == 1432
+    assert session.query(PartVolt).count() == 2310
+    assert session.query(Candidate).count() == 3742
     candidate_extractor.clear_all(split=0)
     assert session.query(Candidate).count() == 0
 
@@ -406,23 +402,23 @@ def test_cand_gen(caplog):
 
     candidate_extractor.apply(docs, split=0, parallelism=PARALLEL)
 
-    assert session.query(PartTemp).count() == 3879
-    assert session.query(PartVolt).count() == 3266
-    assert session.query(Candidate).count() == 7145
+    assert session.query(PartTemp).count() == 1432
+    assert session.query(PartVolt).count() == 1993
+    assert session.query(Candidate).count() == 3425
     assert docs[0].name == "112823"
     assert len(docs[0].parts) == 70
     assert len(docs[0].volts) == 33
-    assert len(docs[0].temps) == 24
+    assert len(docs[0].temps) == 23
 
     # Test that deletion of a Candidate does not delete the Mention
     session.query(PartTemp).delete(synchronize_session="fetch")
     assert session.query(PartTemp).count() == 0
-    assert session.query(Temp).count() == 136
-    assert session.query(Part).count() == 234
+    assert session.query(Temp).count() == 23
+    assert session.query(Part).count() == 70
 
     # Test deletion of Candidate if Mention is deleted
-    assert session.query(PartVolt).count() == 3266
-    assert session.query(Volt).count() == 107
+    assert session.query(PartVolt).count() == 1993
+    assert session.query(Volt).count() == 33
     session.query(Volt).delete(synchronize_session="fetch")
     assert session.query(Volt).count() == 0
     assert session.query(PartVolt).count() == 0
@@ -457,9 +453,9 @@ def test_ngrams(caplog):
     )
     mention_extractor.apply(docs, parallelism=PARALLEL)
 
-    assert session.query(Person).count() == 126
+    assert session.query(Person).count() == 118
     mentions = session.query(Person).all()
-    assert len([x for x in mentions if x.context.get_num_words() == 1]) == 50
+    assert len([x for x in mentions if x.context.get_num_words() == 1]) == 49
     assert len([x for x in mentions if x.context.get_num_words() > 3]) == 0
 
     # Test for unigram exclusion
@@ -468,7 +464,7 @@ def test_ngrams(caplog):
         session, [Person], [person_ngrams], [person_matcher]
     )
     mention_extractor.apply(docs, parallelism=PARALLEL)
-    assert session.query(Person).count() == 76
+    assert session.query(Person).count() == 69
     mentions = session.query(Person).all()
     assert len([x for x in mentions if x.context.get_num_words() == 1]) == 0
     assert len([x for x in mentions if x.context.get_num_words() > 3]) == 0